FirmMonster: Add settings to control fryhtan attack rate.
The alternate attack rate is difficult to test automatically since the AI has a poor response to fryhtans. Effectively the rate is 1/3 against AI. Also, when you reduce the number of lairs, the game gets easier, since the fryhtans attack less often.

Went to a flat random rate like the original game had by default because:
1. AI starts to struggle at the effective 1/3 rate of the curve.
2. Most games don't have that many lairs, and killing a few is easy enough that it doesn't matter anymore, since it will go up towards 1/6 probably quickly. Flat rate will always be the same, keeping the difficulty the same.
3. Flat rate is easier to configure from user perspective.

The curve might make fryhtan fest like maps quite interesting, which is what it really was programmed to. So user option it is. Went up to 1/4 rate to make AI more competitive. Though probably not much difference between 1/4 and 1/6 anyway.