MultiPlayer: Use larger receive buffer.
The game seems to use one large datagram to send the entire queue of a single frame. The original size of 500, and even the larger 800 is too small. We shouldn't be worried about fragmenting, but even if we do, this should be handled in the low level network protocol and we ought not to worry about it. Most frames will likely fit in a datagram. It's probably the crc frames that get pretty large, after a while.

The enet code will now use a generous size up to typical ethernet limits of today, and allow even larger with fragmentation.