Unit: Update attack mode when line of sight is broken in ranged attack.
Some attacks cycle between different modes. The cycling occurs at the end of an animation (frame_count). If the target moves after a bullet is generated, but before the animation is completed, then the cycling of attack mode does not happen as the animation is interrupted by stop(). Later the unit may reacquire the target, and repeat the same attack again without any additional checking. This is would not be much of problem except the attack may no longer be valid due to no longer having enough power points for the same attack's power-level. This can happen with the Egyptian ranged attack, which could allow for more flame arrows. The game authors clearly did not intend to allow powered attacks to occurred on depleted power points.

When the unit breaks the line of sight, update the attack mode that would have happened if the movement didn't interrupt the repeat attack. Since the unit has to move anyway, and this is not a typical repeat action, the best choice is to reset to the best attack available (don't just cycle), which would be the same thing done if a new target is acquired.

The code handling the choosing and cycling of attack modes not pretty. 7k2 tried to clean it up, but the same hole in logic might be there.