UnitCaravan/Marine: Prioritize in-firm processing over stop changes.
Some of the stop change code will modify the journey status immediately to go to the new stop. This is wrong when allowed when in the firm, because the caravan will try to execute moves, which causes attempts to path from "off-map", and may lead to a crash in an extreme case. Changing stops when in the firm also is related to the caravans getting stuck in the firm, because the journey status changed, and unit processing will no longer restore the caravan to the map at the end of the waiting period.

The more correct thing to do is to prioritize the in firm count down, and leave the next stop processing to happen, as it does anyway, when the caravan reappears.

7k2 may have the same issue.