Unit: Disable cycling of ranged attack when target moves and stays in range.
This cycling does not make sense. The game should allow the attack to finish, and cycle then. What ends up happening, for example, if the target keeps moving, an egyptian will keep choosing fire arrows at first, because a unit can move two times between arrows. For example:

1) launch power arrow
2) target move - cycle to normal arrow
3) target move - cycle to power arrow
4) launch power arrow
5) repeat at step (2) until drained

After drained, the process then becomes:
1) launch normal arrow
2) target move - cycle to power arrow if enough power
3) target move - cycle to normal arrow
4) launch normal arrow
5) repeat at step (2) until target dead

This in effect neutralizes the ability for the egyptian attack from being powered on next attack if triggered at the right moment (drained).

This only affects ranged attackers. The only other one is the chinese dragon -- it's not powered, but cycling it too quickly doesn't make sense either.

The fix should be to only cycle on end of animation, and not to interrupt it.

7k2 seems to be better here.