Unit: Allow ai to interrupt orders to help teammate.
The function "ask_team_help_attack" doesn't help the ai in practice, since the ai never lets a unit idle that is on a mission. Once the target is destroyed, it's immediately commanded back to base. This change allows the team to help those still fighting finish the battle within reason, which will reduce losses to the ai players.

This adds an ai-only condition to return to battle team members to help more that have otherwise finished their task.