ZoomMatrix: Check if bullet is in the zoom rectangle before accessing location.
Some bullets are offset enough from the generation point that they are in a different location. If this happens on the edge of the map, an invalid location is referenced in is_shealth(). (Note: is_shealth is a typo, fixed in 7k2 to is_stealth)

Tornadoes have the same problem, but in 7kaa the devs disabled the call to is_shealth in a patch, making them visible in the fog (a new bug). In 7k2, they fixed tornado improper visibility by checking against the world bounds before is_stealth. However, a better fix is just to reverse the order of checking the zoom bounds against is_shealth (done anyway and supersedes the world bound check).

7k2 could have an issue with certain sprite bullets, although all sprites were redone, so they could have eliminated any potential issue that way.